Updated includes of renamed platform header files
