+++ features/TESTS/filesystem/fat_filesystem/main.cpp
@@ -21,7 +21,7 @@
 #include "HeapBlockDevice.h"
 #include "FATFileSystem.h"
 #include <stdlib.h>
-#include "retarget.h"
+#include "mbed_retarget.h"
 
 using namespace utest::v1;
