+++ platform/platform.h
@@ -24,8 +24,8 @@
 #include <cstdio>
 #include <cstring>
 
-#include "platform/retarget.h"
-#include "platform/toolchain.h"
+#include "platform/mbed_retarget.h"
+#include "platform/mbed_toolchain.h"
 #include "device.h"
 #include "PinNames.h"
 #include "PeripheralNames.h"
